backend/requirements.txt
flask===2.2.5
werkzeug==2.2.3

# Applicatoin server for both development and production
gunicorn==20.1.0

# Testing and static anaylsis
pytest==7.4.0
pytest_cov==4.1.0
flake8==5.0.4

# CLI
click==8.1.3

# Data and workers
redis==4.6.0
celery==5.2.4

kombu==4.6.11
vine==1.3.0

# Forms
Flask_WTF==1.1.1

# Extentions
Flask-Mail==0.9.1

python-dotenv